BGS Neural Network Model/requirements.txt
opencv-python~=4.4.0.46
torchvision~=0.8.2
numpy~=1.19.2
scikit-image~=0.17.2
matplotlib~=3.3.2
pillow~=8.0.1
tqdm~=4.50.2